Move Action source files into 'action' directory
